Add actual file upload to db and sign up folder
Now when a user uploads a file they must choose a folder for the file
and the file's information is loaded to database. The file itself is not
sent, instead the url of the location which in the future will be a cdn
link. Also made it so a 'Home' folder is created when a user signs up.